Add check_svnbot plugin
This plugin checks two things:
- Failures of svnbot to handle auth requests done by nginx (svnbot
crashed, or it does not respond etc). This is a statistic kept by nginx
lua code.
- How long ago public signing keys of MxID3 were refreshed, to spot any
failures doing so. Hardcoded values are to warn if it failed for 1 day,
and to go critical if it fails again next day. This is a statistic that
is queried from svnbot directly, also checking if it still can handle
requests.

Command line params are the port numbers for nginx stats and svnbot.