debian/control: remove Suggests
The python-mx-btrfs package does not exist for years any more already,
and python-psycopg2 is a python2 thing that does not exist any more in
the new Debian release.

Just remove all of it, there's a whole bunch of different dependencies
that vary per plugin, we usually fix that using puppet instead.